refactor: HeaderLeftCol setSearchVariant 관련 로직 리팩토링
